+++ src/models/Settings.php
@@ -14,7 +14,13 @@ use craft\base\Model;
 class Settings extends Model
 {
     /**
-     * @var bool
+     * @var string|null
      */
-    public bool $editorCloudApiKey = false;
+    public ?string $editorCloudApiKey = null;
+
+    /**
+     * @var bool
+     * @since 1.3.0
+     */
+    public bool $enablePremiumPlugins = false;
 }